+++ runner/src/Storage/VariableArithmetic.h
@@ -12,7 +12,7 @@ class VariableArithmetic : public Variable {
 
   virtual char get_char() const = 0;
   virtual int get_int() const = 0;
-  virtual long get_long() const = 0;
+  virtual long long get_long() const = 0;
   virtual float get_float() const = 0;
   virtual double get_double() const = 0;
 };
